Commit 03/06/2019

+++ rest/src/main/java/usc/choiceanalyst/controller/ControladorEstablecimiento.java
@@ -13,7 +13,6 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 import usc.choiceanalyst.model.ModeloEstablecimiento;
 import usc.choiceanalyst.model.ModeloExperimento;
-import usc.choiceanalyst.model.ModeloUsuario;
 import usc.choiceanalyst.model.auxiliar.Menu;
 import usc.choiceanalyst.repository.RepositorioEstablecimiento;
 import usc.choiceanalyst.repository.RepositorioExperiencia;
@@ -79,19 +78,6 @@ public class ControladorEstablecimiento {
 
         return ResponseEntity.ok(dbEstablecimiento.findAll(Example.of(establecimiento, ExampleMatcher.matching().withIgnoreCase()), Sort.by(Sort.Order.asc("idEstablecimiento"))));
     }
-
-/*    @PreAuthorize("hasRole('CLIENTE')")
-    @GetMapping(
-            produces = {MediaType.APPLICATION_JSON_UTF8_VALUE, MediaType.APPLICATION_XML_VALUE}
-    )
-    public ResponseEntity<Collection<ModeloEstablecimiento>> getAllEstablecimientos(@RequestParam(value = "idAdministrador", defaultValue = "") String idAdministrador) {
-        ModeloEstablecimiento establecimiento = new ModeloEstablecimiento();
-        if (!idAdministrador.isEmpty()) {
-            establecimiento.setIdAdministrador(idAdministrador);
-        }
-        Collection<ModeloEstablecimiento> establecimientos = dbEstablecimiento.findAll(Example.of(establecimiento));
-        return ResponseEntity.ok(establecimientos);
-    }*/
 
 
     @PreAuthorize("hasRole('ADMINISTRADOR')")
@@ -172,21 +158,25 @@ public class ControladorEstablecimiento {
             produces = {MediaType.APPLICATION_JSON_UTF8_VALUE, MediaType.APPLICATION_XML_VALUE}
     )
     public ResponseEntity deleteMenu(@PathVariable("idEstablecimiento") String idEstablecimiento, @PathVariable("idMenu") String idMenu) {
-        if (dbEstablecimiento.existsByIdEstablecimiento(idEstablecimiento)) {
-            Optional<ModeloEstablecimiento> establecimiento = dbEstablecimiento.findByIdEstablecimiento(idEstablecimiento);
-            if (establecimiento.get().getIdAdministrador().equals(SecurityContextHolder.getContext().getAuthentication().getPrincipal().toString())) {
-                for (int i = 0; i < establecimiento.get().getMenus().size(); i++) {
-                    if (establecimiento.get().getMenus().get(i).getIdMenu().equals(idMenu)) {
-                        establecimiento.get().getMenus().remove(i);
+        if (!dbExperiencia.existsByIdEstablecimiento(idEstablecimiento)) {
+            if (dbEstablecimiento.existsByIdEstablecimiento(idEstablecimiento)) {
+                Optional<ModeloEstablecimiento> establecimiento = dbEstablecimiento.findByIdEstablecimiento(idEstablecimiento);
+                if (establecimiento.get().getIdAdministrador().equals(SecurityContextHolder.getContext().getAuthentication().getPrincipal().toString())) {
+                    for (int i = 0; i < establecimiento.get().getMenus().size(); i++) {
+                        if (establecimiento.get().getMenus().get(i).getIdMenu().equals(idMenu)) {
+                            establecimiento.get().getMenus().remove(i);
+                        }
                     }
+                    dbEstablecimiento.save(establecimiento.get());
+                    return ResponseEntity.noContent().build();
+                } else {
+                    return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
                 }
-                dbEstablecimiento.save(establecimiento.get());
-                return ResponseEntity.noContent().build();
             } else {
-                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+                return ResponseEntity.notFound().build();
             }
         } else {
-            return ResponseEntity.notFound().build();
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
         }
     }
 
@@ -194,16 +184,20 @@ public class ControladorEstablecimiento {
     @PreAuthorize("hasRole('ADMINISTRADOR')")
     @DeleteMapping(path = "/{idEstablecimiento}")
     public ResponseEntity deleteEstablecimiento(@PathVariable("idEstablecimiento") String idEstablecimiento) {
-        if (!dbEstablecimiento.existsByIdEstablecimiento(idEstablecimiento)) {
-            return ResponseEntity.notFound().build();
-        } else {
-            ModeloEstablecimiento establecimiento = dbEstablecimiento.findByIdEstablecimiento(idEstablecimiento).get();
-            if (establecimiento.getIdAdministrador().equals(SecurityContextHolder.getContext().getAuthentication().getPrincipal().toString())) {
-                dbEstablecimiento.deleteByIdEstablecimiento(idEstablecimiento);
-                return ResponseEntity.noContent().build();
-            } else {
+        if (!dbExperiencia.existsByIdEstablecimiento(idEstablecimiento)) {
+            if (!dbEstablecimiento.existsByIdEstablecimiento(idEstablecimiento)) {
                 return ResponseEntity.notFound().build();
+            } else {
+                ModeloEstablecimiento establecimiento = dbEstablecimiento.findByIdEstablecimiento(idEstablecimiento).get();
+                if (establecimiento.getIdAdministrador().equals(SecurityContextHolder.getContext().getAuthentication().getPrincipal().toString())) {
+                    dbEstablecimiento.deleteByIdEstablecimiento(idEstablecimiento);
+                    return ResponseEntity.noContent().build();
+                } else {
+                    return ResponseEntity.notFound().build();
+                }
             }
+        } else {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
         }
     }
 
@@ -238,28 +232,33 @@ public class ControladorEstablecimiento {
     )
 
     public ResponseEntity modifyMenu(@RequestBody Menu menu, @PathVariable("idEstablecimiento") String idEstablecimiento, @PathVariable("idMenu") String idMenu) {
-        if (!dbEstablecimiento.existsByIdEstablecimiento(idEstablecimiento)) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
-        } else {
-            ModeloEstablecimiento establecimiento = dbEstablecimiento.findByIdEstablecimiento(idEstablecimiento).get();
-            if (establecimiento.getIdAdministrador().equals(SecurityContextHolder.getContext().getAuthentication().getPrincipal().toString())) {
-                if (!idMenu.equals(menu.getIdMenu())) {
-                    for (int i = 0; i < establecimiento.getMenus().size(); i++) {
-                        if (establecimiento.getMenus().get(i).getIdMenu().equals(menu.getIdMenu())) {
-                            return ResponseEntity.status(HttpStatus.CONFLICT).build();
+        if (!dbExperiencia.existsByIdEstablecimiento(idEstablecimiento)) {
+
+            if (!dbEstablecimiento.existsByIdEstablecimiento(idEstablecimiento)) {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+            } else {
+                ModeloEstablecimiento establecimiento = dbEstablecimiento.findByIdEstablecimiento(idEstablecimiento).get();
+                if (establecimiento.getIdAdministrador().equals(SecurityContextHolder.getContext().getAuthentication().getPrincipal().toString())) {
+                    if (!idMenu.equals(menu.getIdMenu())) {
+                        for (int i = 0; i < establecimiento.getMenus().size(); i++) {
+                            if (establecimiento.getMenus().get(i).getIdMenu().equals(menu.getIdMenu())) {
+                                return ResponseEntity.status(HttpStatus.CONFLICT).build();
+                            }
                         }
                     }
-                }
-                for (int i = 0; i < establecimiento.getMenus().size(); i++) {
-                    if (establecimiento.getMenus().get(i).getIdMenu().equals(idMenu)) {
-                        establecimiento.getMenus().set(i, menu);
+                    for (int i = 0; i < establecimiento.getMenus().size(); i++) {
+                        if (establecimiento.getMenus().get(i).getIdMenu().equals(idMenu)) {
+                            establecimiento.getMenus().set(i, menu);
+                        }
                     }
+                } else {
+                    return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
                 }
-            } else {
-                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+                dbEstablecimiento.save(establecimiento);
+                return ResponseEntity.ok().body(menu);
             }
-            dbEstablecimiento.save(establecimiento);
-            return ResponseEntity.ok().body(menu);
+        } else {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
         }
     }
 
@@ -271,38 +270,42 @@ public class ControladorEstablecimiento {
     )
 
     public ResponseEntity modifyEstablecimiento(@RequestBody ModeloEstablecimiento establecimiento, @PathVariable("idEstablecimiento") String idEstablecimiento) {
-        ModeloEstablecimiento establecimientoExistente;
-        if (!dbEstablecimiento.existsByIdEstablecimiento(idEstablecimiento)) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
-        } else {
-            establecimientoExistente = dbEstablecimiento.findByIdEstablecimiento(idEstablecimiento).get();
-            if (establecimientoExistente.getIdAdministrador().equals(SecurityContextHolder.getContext().getAuthentication().getPrincipal().toString())) {
-                if (!idEstablecimiento.equals(establecimiento.getIdEstablecimiento())) {
-                    if (dbEstablecimiento.existsByIdEstablecimiento(establecimiento.getIdEstablecimiento())) {
-                        return ResponseEntity.status(HttpStatus.CONFLICT).build();
-                    } else {
-                        establecimientoExistente.setIdEstablecimiento(establecimiento.getIdEstablecimiento());
-                        List<ModeloExperimento> experimentos = dbExperimento.findByIdAdministradorAndIdsEstablecimientosContains(establecimientoExistente.getIdAdministrador(), idEstablecimiento);
-                        for (int i = 0; i < experimentos.size(); i++) {
-                            for (int j = 0; j < experimentos.get(i).getIdsEstablecimientos().size(); j++) {
-                                if (experimentos.get(i).getIdsEstablecimientos().get(j).equals(idEstablecimiento)) {
-                                    experimentos.get(i).getIdsEstablecimientos().set(j, establecimiento.getIdEstablecimiento());
-                                    dbExperimento.save(experimentos.get(i));
+        if (!dbExperiencia.existsByIdEstablecimiento(idEstablecimiento)) {
+            ModeloEstablecimiento establecimientoExistente;
+            if (!dbEstablecimiento.existsByIdEstablecimiento(idEstablecimiento)) {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+            } else {
+                establecimientoExistente = dbEstablecimiento.findByIdEstablecimiento(idEstablecimiento).get();
+                if (establecimientoExistente.getIdAdministrador().equals(SecurityContextHolder.getContext().getAuthentication().getPrincipal().toString())) {
+                    if (!idEstablecimiento.equals(establecimiento.getIdEstablecimiento())) {
+                        if (dbEstablecimiento.existsByIdEstablecimiento(establecimiento.getIdEstablecimiento())) {
+                            return ResponseEntity.status(HttpStatus.CONFLICT).build();
+                        } else {
+                            establecimientoExistente.setIdEstablecimiento(establecimiento.getIdEstablecimiento());
+                            List<ModeloExperimento> experimentos = dbExperimento.findByIdAdministradorAndIdsEstablecimientosContains(establecimientoExistente.getIdAdministrador(), idEstablecimiento);
+                            for (int i = 0; i < experimentos.size(); i++) {
+                                for (int j = 0; j < experimentos.get(i).getIdsEstablecimientos().size(); j++) {
+                                    if (experimentos.get(i).getIdsEstablecimientos().get(j).equals(idEstablecimiento)) {
+                                        experimentos.get(i).getIdsEstablecimientos().set(j, establecimiento.getIdEstablecimiento());
+                                        dbExperimento.save(experimentos.get(i));
+                                    }
                                 }
                             }
+                            dbEstablecimiento.deleteByIdEstablecimiento(idEstablecimiento);
                         }
-                        dbEstablecimiento.deleteByIdEstablecimiento(idEstablecimiento);
                     }
-                }
-                establecimientoExistente.setLocalizacionEstablecimiento(establecimiento.getLocalizacionEstablecimiento());
-                establecimientoExistente.setNombreEstablecimiento(establecimiento.getNombreEstablecimiento());
-                establecimientoExistente.setTipoEstablecimiento(establecimiento.getTipoEstablecimiento());
-                dbEstablecimiento.save(establecimientoExistente);
+                    establecimientoExistente.setLocalizacionEstablecimiento(establecimiento.getLocalizacionEstablecimiento());
+                    establecimientoExistente.setNombreEstablecimiento(establecimiento.getNombreEstablecimiento());
+                    establecimientoExistente.setTipoEstablecimiento(establecimiento.getTipoEstablecimiento());
+                    dbEstablecimiento.save(establecimientoExistente);
 
-                return ResponseEntity.ok().body(establecimientoExistente);
-            } else {
-                return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+                    return ResponseEntity.ok().body(establecimientoExistente);
+                } else {
+                    return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+                }
             }
+        } else {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
         }
     }
